requested changes & refactor to omnipaxos_storage - Split memory.rs into persistent_storage.rs and memory_storage.rs - Error messages in serialization, deserialization, reads and writes to storage - Private wrapper structs - Refactor, optimize getters and setters - test_conf and util merged into utils.rs
