+++ omnipaxos_core/tests/config/test.conf
@@ -5,7 +5,7 @@ ble_test {
     ble_hb_delay: 5
     increment_delay: 2
     num_elections: 3
-    storage_type: Persistent
+    storage_type: persistent
 }
 
 consensus_test {
@@ -15,7 +15,7 @@ consensus_test {
     ble_hb_delay: 5
     increment_delay: 2
     num_proposals: 20
-    storage_type: Persistent
+    storage_type: persistent
 }
 
 proposal_test {
@@ -24,7 +24,7 @@ proposal_test {
     num_nodes: 7
     ble_hb_delay: 5
     increment_delay: 2
-    storage_type: Persistent
+    storage_type: persistent
 }
 
 gc_test {
@@ -35,5 +35,5 @@ gc_test {
     increment_delay: 2
     num_proposals: 30
     gc_idx: 10
-    storage_type: Persistent
+    storage_type: persistent
 }
