--- omnipaxos_core/tests/utils/mod.rs
@@ -1,6 +1,0 @@
-// Submodule for util and test config
-
-// Configuration for tests
-pub mod test_config;
-// Test environment creation and utilities
-pub mod util;
