+++ omnipaxos_storage/src/lib.rs
@@ -1,7 +1,7 @@
-//! An library of storage implementations for sequencePaxos, provides
-//! in-memory storage implementation with fast read and writes, and 
-//! an persistent storage implementation with persistence for the 
-//! replica state and the log through disk storage.
+//! An library of storage implementations for sequencePaxos
 
 //#![deny(missing_docs)]
-pub mod memory;
+/// an in-memory storage implementation with fast read and writes
+pub mod memory_storage;
+/// an on-disk storage implementation with persistence for the replica state and the log.
+pub mod persistent_storage;
