+++ rustfmt.toml
@@ -49,6 +49,8 @@ skip_children = false
 hide_parse_errors = false
 error_on_line_overflow = false
 error_on_unformatted = false
+report_todo = "Never"
+report_fixme = "Always"
 ignore = []
 emit_mode = "Files"
 make_backup = false
